Move change_set and change_set_workflow out of Valhalla
