+++ .rubocop.yml
@@ -34,7 +34,7 @@ Metrics/BlockLength:
     - 'db/schema.rb'
     - 'lib/tasks/dev.rake'
     - 'app/services/pdf_generator/cover_page_generator.rb'
-    - 'valhalla/app/change_sets/valhalla/change_set_workflow.rb'
+    - 'app/change_sets/change_set_workflow.rb'
     - 'lib/tasks/bulk.rake'
     - 'lib/tasks/lae.rake'
     - 'app/change_sets/simple_resource_change_set.rb'
