+++ app/change_sets/ephemera_box_change_set.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-class EphemeraBoxChangeSet < Valhalla::ChangeSet
+class EphemeraBoxChangeSet < ChangeSet
   apply_workflow(WorkflowRegistry.workflow_for(EphemeraBox))
   validates :barcode, :box_number, :visibility, :rights_statement, presence: true
