+++ app/change_sets/vector_resource_change_set.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-class VectorResourceChangeSet < Valhalla::ChangeSet
+class VectorResourceChangeSet < ChangeSet
   apply_workflow(WorkflowRegistry.workflow_for(VectorResource))
   delegate :human_readable_type, to: :model
